[FEAT] 검색 기록 기반 검색
